chore: move the ListOfHikeRoutesViewModel
The ListOfHikeRoutesViewModel is more suitable to be in model.route than model.map, since it doesn't provide any data for the map itself, just data to show on top of it
